refactor: Table 관련된 사용하지 않는 컴포넌트 삭제, TableBox 이름 Table로 변경
