+++ components/store/infoTable/Table/Table.styled.tsx
@@ -1,40 +1,31 @@
 import styled from "@emotion/styled";
-import { css } from "@emotion/react";
 
-interface ICellProps {
-  option: string;
-}
+export const Key = styled.div`
+  background-color: ${({ theme }) => theme.colors.grey[200]};
+`;
 
-export const Container = styled.div`
-  width: 32rem;
-  display: flex;
-  margin: 0 auto;
-  border-top: 1px solid ${({ theme }) => theme.colors.grey[600]};
+export const Value = styled.div`
+  background-color: ${({ theme }) => theme.colors.grey[100]};
 
-  :last-child {
-    border-bottom: 1px solid ${({ theme }) => theme.colors.grey[600]};
+  > p {
+    color: ${({ theme }) => theme.colors.grey[800]};
+    line-height: 150%;
   }
 `;
 
-export const Cell = styled.div<ICellProps>`
-  ${({ option, theme }) => {
-    return option === "left"
-      ? css`
-          width: 40%;
-          font-size: ${theme.fontSizes[14]};
-          border-right: 1px solid ${theme.colors.grey[600]};
-          font-weight: 500;
-          width: 50%;
-          line-height: 4.1rem;
-          padding-left: 1.6rem;
-          background-color: ${theme.colors.navy[100]};
-        `
-      : css`
-          width: 60%;
-          font-size: ${theme.fontSizes[14]};
-          font-weight: 500;
-          padding: 1.4rem 1.6rem;
-          /* border-right: 1px solid ${theme.colors.grey[600]}; */
-        `;
-  }}
+export const Grid = styled.div<{ option: "notice" | "info" }>`
+  display: grid;
+  grid-gap: 0.06rem;
+  grid-template-columns: [sidebar] 2fr [content] 3fr;
+  border-top: 1.2px solid ${({ theme }) => theme.colors.grey[500]};
+  border-bottom: 1.2px solid ${({ theme }) => theme.colors.grey[500]};
+  background-color: ${({ theme }) => theme.colors.grey[400]};
+
+  > * {
+    padding: 1.4rem 1.6rem;
+  }
+
+  ${Value} {
+    text-align: ${({ option }) => (option === "notice" ? "center" : "left")};
+  }
 `;
