+++ components/store/infoTable/Table/Table.tsx
@@ -1,13 +1,11 @@
 import React from "react";
+import { IKeyValue, ITime } from "@/api/types/shared";
+import Text from "@/components/shared/Text/Text";
 import * as S from "./Table.styled";
 
-interface IContents {
-  option: string;
-  data: ({ day: string; hour: string } | { key: string; value: string })[];
-}
-
-interface IEl {
-  [key: string]: string;
+interface IInfo {
+  option: "notice" | "info";
+  data: (IKeyValue | ITime)[];
 }
 
 interface IOptions {
@@ -25,18 +23,25 @@ const optionList: IOptions = {
   },
 };
 
-// TODO: props에서 option 제거
-export default function Table({ option, data }: IContents) {
+export default function TableBox({ option, data }: IInfo) {
   const { key, value } = optionList[option];
 
   return (
-    <>
-      {data.map((el: IEl) => (
-        <S.Container key={el[key]}>
-          <S.Cell option="left">{el[key]}</S.Cell>
-          <S.Cell option="right">{el[value]}</S.Cell>
-        </S.Container>
+    <S.Grid option={option}>
+      {data.map(tableItem => (
+        <React.Fragment key={tableItem[key]}>
+          <S.Key>
+            <Text weight={500} as="h3">
+              {tableItem[key]}
+            </Text>
+          </S.Key>
+          <S.Value>
+            <Text weight={400} as="p">
+              {tableItem[value]}
+            </Text>
+          </S.Value>
+        </React.Fragment>
       ))}
-    </>
+    </S.Grid>
   );
 }
